代码重构

+++ scene/element/bird/bird.js
@@ -9,6 +9,7 @@ class Bird extends Entity {
         this.y = 0
         this.h = 40
         this.w = 40
+        // 只要小鸟没有掉到最下面，都算是在飞
         this.flying = true
         this.dropSpeed = 5
         this.jumpSpeed = -10
@@ -16,10 +17,13 @@ class Bird extends Entity {
     }
 
     update() {
-        this.drop()
+        if (this.flying) {
+            this.drop()
+        }
     }
 
     drop() {
+        // 调整角度使头向下
         this.rotation = 45
         this.flying = this.y < window.height - this.h
         if (this.flying) {
@@ -28,16 +32,16 @@ class Bird extends Entity {
     }
 
     jump() {
-        if (!this.dead) {
-            this.rotation = -45
-            if (this.y > 0) {
-                this.y += this.jumpSpeed
-            }
+        if (this.dead) {
+            return
+        }
+        // 调整角度使头向上
+        this.rotation = -45
+        var touchTop = this.y <= 0
+        if (!touchTop) {
+            this.y += this.jumpSpeed
         }
     }
 
-    handleCollide() {
-        this.dead = true
-    }
 
 }
